MorphingPortableReader is implemented and tested
